ssh-agent Signer
Worked example of how to use an `ssh-agent(2)` as a `Signer` impl.

There are some limitations of the `thrussh-keys` crate, however:

* The signature of `AgentClient::sign_request_signature` is probably
  wrong, which leads to the requirement of setting the `SSH_AUTH_SOCK`
  environment variable
* The constraints types are either wrong, or the standard distro agent
  doesn't understand them (or not all of them, eg. KeyLifetime)
* It is unfortunate that `sodiumoxide` and `tokio` get pulled in.
  Perhaps Pierre can be convinced of the idea to abstract over the former,
  at least.
